Use a default attribute instead of condition
This ended up being a quick addition of an attribute without a default
value, and instead could have always just used the attribute, and set
the default like so.

End result of this is indentical to the current code, with the addition
of the attribute having a value regardless.

See cca0272 for details on the previous change