Correct how install_method is implemented
As per Exhibitor’s docs:

https://github.com/Netflix/exhibitor/wiki/Building-Exhibitor

there are two ways to assemble, Maven and Gradle.

This cookbook has had the Gradle method implemented for some time, but
seemed to be under the impression that the JAR could be downloaded and
used as-is. Perhaps some have pre-built JARs, and that functionality can
be added back, but for now, to avoid confusion, this is how this will be
implemented.

Maven functionality TBD.